completing openvpn conf

+++ bin/openvpn/openvpn_server.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
 cp -R /usr/share/doc/openvpn/examples/easy-rsa/ /etc/openvpn
-cat ./whichopensslcnf > /etc/openvpn/easy-rsa/2.0/whichopensslcnf
+cat $DIR/whichopensslcnf > /etc/openvpn/easy-rsa/2.0/whichopensslcnf
 
 cd /etc/openvpn/easy-rsa/2.0/
 
 # Make the variables
 source ./vars
+
+source $DIR/openvpn_vars.sh
+
 source ./clean-all
 
 # Create the Certificate Authority
@@ -14,6 +19,9 @@ source ./build-ca
 
 # Create a certificate/key for your server
 source ./build-key-server server
+
+# Diffie Hellman parameters must be generated for the OpenVPN server
+source ./build-dh
 
 cd keys
 
@@ -24,13 +32,12 @@ cd /usr/share/doc/openvpn/examples/sample-config-files
 
 [ -e 'server.conf.gz' ] && gunzip -d server.conf.gz
 
-# Backup the previous server.conf
-[ -e '/etc/openvpn/server.conf' ] && \
-  echo "Backing up server.conf to server.conf.bck"
-  mv /etc/openvpn/server.conf /etc/openvpn/server.conf.bck
-
 # Copy the sample server configuration to /etc/openvpn
-cp server.conf /etc/openvpn/
+if [[ -e '/etc/openvpn/server.conf' ]]; then
+  echo "server.conf exists. Not overwritting"
+else
+  cp server.conf /etc/openvpn/
+fi
 
 /etc/init.d/openvpn stop
 /etc/init.d/openvpn start
